- manage customer data and loyalty data on dashboard - code optimization

+++ app/src/main/java/com/android/pos/data/entities/LoyaltyProgramsModel.kt
@@ -24,7 +24,7 @@ data class LoyaltyProgramsModel(
     @SerializedName("name")
     val name: String,
     @SerializedName("reward_point")
-    val rewardPoint: Double,
+    val rewardPoint: Int,
     @SerializedName("reward_type")
     val rewardType: String,
     @SerializedName("updated_at")
